moved some shared functions to utils.js
