+++ featured.js
@@ -1,4 +1,6 @@
-import { Auth, sortingFunctions } from "./auth.js";
+import { Auth } from "./auth.js";
+import { sortingFunctions, getHeartPath, populateModalData } from "./utils.js";
+
 
 const auth = await Auth.setUpAuth();
 
@@ -52,7 +54,7 @@ async function renderFeaturedPlaylistList() {
     document.getElementById("randomFeaturedPlaylistTitle").innerText = data.playlists[randomPlaylistIndex].playlist_name;
     document.getElementById("randomFeaturedPlaylistCreator").innerText = data.playlists[randomPlaylistIndex].playlist_creator;
     document.getElementById("randomFeaturedPlaylistLikesCount").innerText = data.playlists[randomPlaylistIndex].liked_users.length;
-    const randomFeaturedHeartPath = auth.getHeartPath(data.playlists[randomPlaylistIndex].liked_users, user.email);
+    const randomFeaturedHeartPath = getHeartPath(data.playlists[randomPlaylistIndex].liked_users, user.email);
     document.getElementById("randomFeaturedPlaylistItemHeart").src = randomFeaturedHeartPath;
 
     document.getElementById("randomFeaturedPlaylistLikeContainer").addEventListener("click", function (e) {
@@ -67,7 +69,7 @@ async function renderFeaturedPlaylistList() {
     playlistsContainer.innerHTML = "";
     // rendering the playlists
     data.playlists.forEach((playlist) => {
-        const currentHeartPath = auth.getHeartPath(playlist.liked_users, user.email);
+        const currentHeartPath = getHeartPath(playlist.liked_users, user.email);
 
         let playlistItem = document.createElement("li"); // Create a <li> element
         playlistItem.innerHTML = `
@@ -266,31 +268,7 @@ async function openModal(playlistToOpen) {
         }
     }
 
-    //TODO: switch to binary search
-    // console.log(data.playlists.filter((playlist) => {
-    //     return playlist.playlistID == playlistIDToGet;
-    // })[0].playlist_name)
-    // console.log(playlistIDToGet + "  " + "");
-    // playlistToOpen = getPlaylistByID(playlistIDToGet);
-    //  = data.playlists.filter((playlist) => {
-    //     return playlist.playlistID == playlistIDToGet;
-    // })[0];
-    console.log(playlistToOpen);
-    document.getElementById("playlistModalName").innerText = playlistToOpen.playlist_name;
-    document.getElementById("playlistModalImage").src = playlistToOpen.playlist_art;
-    document.getElementById("playlistModalCreatorName").innerText = playlistToOpen.playlist_creator;
-    document.getElementById("playlistModalLikesCount").innerText = playlistToOpen.liked_users.length;
-
-    const currentHeartPath = auth.getHeartPath(playlistToOpen.liked_users, user.email);
-    // alert(currentHeartPath)
-    const heartModalElement = document.getElementById("playlistModalHeart");
-    heartModalElement.src = currentHeartPath;
-
-    // Clones the node to remove the previous event listners
-    // https://stackoverflow.com/questions/9251837/how-to-remove-all-listeners-in-an-element
-    let old_element = document.getElementById("playlistModalLikesContainer");
-    let new_element = old_element.cloneNode(true);
-    old_element.parentNode.replaceChild(new_element, old_element);
+    populateModalData(playlistToOpen, user.email);
 
     renderSongList(playlistToOpen);
