+++ myPlaylists.html
@@ -10,7 +10,7 @@
 
     <link rel="stylesheet" href="style.css">
     <!-- <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js@2"></script> -->
-    
+
 </head>
 
 <body>
@@ -21,7 +21,7 @@
     <nav>
         <div id="userInfoAndButtons">
             <div>
-                
+
                 <a href="/">Home</a>
                 <button id="loadExampleDataButton">Load example data</button>
                 <select name="sortSelecter" id="sortSelecter">
@@ -48,10 +48,10 @@
                 <button id="signOut">Sign Out</button>
 
             </div>
-            
+
 
         </div>
-        
+
     </nav>
 
     <!-- Modal Structure -->
@@ -167,7 +167,7 @@
                         <input type="text" id="addSongDuration" name="addSongDuration"  />
                         <label for="addSongDuration">Song song duration</label>
                     </div>
-                    
+
                 </p>
                 <div>
                     <button value="cancel">Cancel</button>
@@ -178,7 +178,7 @@
 
     <main>
 
-        
+
         <!-- Your main content here -->
         <ul id="playlistsContainer" class="playlistCards">
 
@@ -205,7 +205,7 @@
     <!-- <script src="data/data.js"></script> -->
     <!-- <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js@2"></script> -->
 
-    <script type="module" src="script.js"></script>
+    <script type="module" src="myPlaylists.js"></script>
 </body>
 
 </html>
